Fix deps; rename backend; scaffold new backend.
