+++ src/main/java/cloud/filibuster/junit/configuration/FilibusterConfiguration.java
@@ -1,7 +1,7 @@
 package cloud.filibuster.junit.configuration;
 
 import cloud.filibuster.instrumentation.helpers.Networking;
-import cloud.filibuster.junit.server.FilibusterDebugServerBackend;
+import cloud.filibuster.junit.server.backends.FilibusterLocalProcessServerBackend;
 import cloud.filibuster.junit.server.FilibusterServerBackend;
 import com.google.errorprone.annotations.CanIgnoreReturnValue;
 import io.micrometer.core.instrument.util.IOUtils;
@@ -111,7 +111,7 @@ public class FilibusterConfiguration {
 
         private String analysisFile;
 
-        private FilibusterServerBackend filibusterServerBackend = new FilibusterDebugServerBackend();
+        private FilibusterServerBackend filibusterServerBackend = new FilibusterLocalProcessServerBackend();
 
         /**
          * Should this configuration use dynamic reduction?
