+++ src/main/java/cloud/filibuster/junit/server/backends/FilibusterLocalProcessServerBackend.java
@@ -1,14 +1,15 @@
-package cloud.filibuster.junit.server;
+package cloud.filibuster.junit.server.backends;
 
 import cloud.filibuster.junit.configuration.FilibusterConfiguration;
+import cloud.filibuster.junit.server.FilibusterServerBackend;
 
 import javax.annotation.Nullable;
 import java.io.IOException;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
-public class FilibusterDebugServerBackend implements FilibusterServerBackend {
-    private static final Logger logger = Logger.getLogger(FilibusterDebugServerBackend.class.getName());
+public class FilibusterLocalProcessServerBackend implements FilibusterServerBackend {
+    private static final Logger logger = Logger.getLogger(FilibusterLocalProcessServerBackend.class.getName());
 
     @Nullable
     private static Process filibusterServerProcess;
